refactor : change package name egit -> utils
